requirements.txt
transformers
transformers[sentencepiece]
datasets
sacrebleu 
rouge_score 
py7zr
pandas
nltk
tqdm
PyYAML
matplotlib
torch
notebook
boto3
mypy-boto3-s3
python-box==6.0.2
ensure==1.0.2
fastapi==0.78.0
uvicorn==0.18.3
Jinja2==3.1.2
evaluate

python-dotenv==1.0.1
requests==2.32.3
pytest==8.3.3
pyyaml==6.0.2
python-box==7.2.0
ensure==1.0.4
transformers==4.44.2
torch==2.4.1
pillow==10.4.0

opencv-python
deepface
tf_keras
tk
tensorflow
numpy
ultralytics 
